Update README.md - Added section regarding exercise naming
Later on, when we implement analysis, exercise names will have to be the same other wise when adding them. eg 'barbell curls' and 'bb curls' and 'bb curl' would be stored as 3 different exercises. Long term goal might be a tab complete and pre-built exercise index. Not right now though. I want to keep this simple.